# perf(ext/http): experimental speedups for serve