@bartlomieju 2023-09-13
I suggest we land the changes in bite-sized fashion. Could we start with just the buffer optimization? It will be easier to figure out what (potentially) breaks WPT.